+++ src/util.cpp
@@ -50,7 +50,24 @@ std::string AltaCore::Util::unescape(std::string data) {
   for (size_t i = 0; i < data.length(); i++) {
     auto& character = data[i];
     if (character == '\\' && i + 1 < data.length()) {
-      result += data[i + 1];
+      const char nextChar = data[i + 1];
+      if (nextChar == 't') {
+        result += '\t';
+      } else if (nextChar == 'r') {
+        result += '\r';
+      } else if (nextChar == 'n') {
+        result += '\n';
+      } else if (nextChar == 'x' && i + 3 < data.length()) {
+        const uint8_t upperHalf = hexDigitToDecimal(data[i + 2]);
+        const uint8_t lowerHalf = hexDigitToDecimal(data[i + 3]);
+        result += (char)((upperHalf << 4) | lowerHalf);
+
+        // skip the 2 next characters
+        i++;
+        i++;
+      } else {
+        result += nextChar;
+      }
       i++; // skip the next character
     } else {
       result += character;
@@ -58,4 +75,16 @@ std::string AltaCore::Util::unescape(std::string data) {
   }
 
   return result;
+};
+
+uint8_t AltaCore::Util::hexDigitToDecimal(const char singleDigit) {
+  if (singleDigit >= '0' && singleDigit <= '9') {
+    return singleDigit - 0x30; // according to asciitable.com, '0' starts at 0x30
+  } else if (singleDigit >= 'A' && singleDigit <= 'F') {
+    return singleDigit - 0x41;
+  } else if (singleDigit >= 'a' && singleDigit <= 'f') {
+    return singleDigit - 0x61;
+  } else {
+    throw std::runtime_error(std::string("invalid hex digit: ") + singleDigit);
+  }
 };